moves spotify api lib to resources
makes more sense for spotify to be in resources because that's what we
use to communicate with spotify api
